DBを使用しないmain.goを作成
